points refactored to double, prepared for swiss impl

+++ src/main/java/cz/tournament/control/service/FileGeneratorService.java
@@ -149,7 +149,7 @@ public class FileGeneratorService {
         Collections.sort(evals, new Comparator<EvalClassic>() {
             @Override
             public int compare(EvalClassic p1, EvalClassic p2) {
-                return Integer.compare(p1.points, p2.points);
+                return Double.compare(p1.points, p2.points);
             }
         });
 
@@ -184,9 +184,11 @@ public class FileGeneratorService {
 
     }
 
-    private int countPoints(AllVersusAll tournament, int wins, int loses, int ties) {
+    private Double countPoints(AllVersusAll tournament, int wins, int loses, int ties) {
         return (wins * tournament.getPointsForWinning()) + (ties * tournament.getPointsForTie()) - (loses * tournament.getPointsForLosing());
     }
+    
+   
 
 
 /*    {"Round", "Rival A", ":", "Rival B", "Score A", ":", "Score B"} */
